requirements/base.txt
Django==1.6.4
South==0.8.4
dj-database-url==0.3.0
#psycopg2==2.5.2
docutils==0.11
django-admin-bootstrapped==1.6.4
djangorestframework==2.3.13
django-cors-headers==0.12
gunicorn==18.0
dj-static==0.0.5
static==0.4
wsgiref==0.1.2